Update to .NET 9.0 in YAML workflows
This commit updates the paths for the `AssetRipper.Tpk.ConsoleApp` executable in `engine_assets_tpk.yml` and `type_tree_tpk.yml`. References to the .NET 6.0 version have been replaced with the .NET 9.0 version. The changes include making the new executable executable and running it with the appropriate arguments for generating TPK files from engine assets and type tree dumps, ensuring the latest version is utilized in the workflows.